Ejected logic from JsonBeautify widget
